requirements.txt
flask
psycopg2-binary
boto3
awscli